code update
The final version improved the robustness of the code while ensuring that tasks can be completed smoothly. It added interesting code effects to distinguish my code from others, and also added comments to the code to ensure that the specific structure of the code can be understood at any time.